requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-generativeai
requests>=2.31.0
cryptography>=41.0.0
python-multipart>=0.0.6
ebaysdk>=2.2.0
asgiref>=3.7.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
gunicorn>=20.1.0
google-cloud-secret-manager
google-cloud-storage
google-cloud-logging